YOLO annotation weights to haarcascade annotations
If you have YOLO darknet annotations (txt file) and dataset and you want to train haarcascade model of that dataset and you don't want to annotate the images again then use this script. It will read the images from dataset and its annotated file and by using the annotations, it will crop the specific object and save it in the form of images in the folder that haarcascade will use as positive images.